updated tests

+++ tests/branching_kernels/if_2/if_2_dcmpl-gfx1030.cl
@@ -3,7 +3,7 @@ void if_2(int x, __global int *data, int y)
 {
     uint var2;
     var2 = get_global_id(0);
-    if ((uint)0 == (uint)var2) {
+    if (0 == var2) {
         var2 = get_global_id(1);
     }
     data[var2] = x;
