+++ tests/branching_kernels/if_else_0_labels/if_else_0_labels_hands-gfx.cl
@@ -5,14 +5,14 @@ void if_else_0_labels(int x, __global int *data, int y)
     uint var2;
     uint var4;
     var2 = get_global_id(0);
-    if ((uint)1 != (uint)var2) {
+    if (1 != var2) {
         var4 = get_global_id(1);
         var1 = y;
     }
     else {
         var4 = get_global_id(1);
         var2 = 1;
-        var1 = (uint)(get_global_id(1) * x) - (uint)y;
+        var1 = (get_global_id(1) * x) - (uint)y;
     }
     data[var2] = var1;
     data[var4] = x;
