+++ tests/branching_kernels/if_else_1_label/if_else_1_label_hands-gfx.cl
@@ -5,14 +5,14 @@ void if_else_1_label(int x, __global int *data, int y)
     uint var3;
     uint var5;
     var3 = get_global_id(0);
-    if ((uint)2 != (uint)var3) {
+    if (2 != var3) {
         var1 = get_global_id(0) * y;
         var5 = get_global_id(1);
     }
     else {
         var5 = get_global_id(1);
         var3 = 2;
-        var1 = (uint)(get_global_id(1) * x) - (uint)y;
+        var1 = (get_global_id(1) * x) - (uint)y;
     }
     data[var3] = var1;
     data[var5] = x;
