+++ tests/branching_kernels/if_else_2_labels/if_else_2_labels_dcmpl-gfx1030.cl
@@ -5,14 +5,14 @@ void if_else_2_labels(int x, __global int *data, int y)
     uint var3;
     uint var5;
     var3 = get_global_id(0);
-    if ((uint)1 != (uint)var3) {
+    if (1 != var3) {
         var5 = get_global_id(1);
         var2 = y * x;
     }
     else {
         var5 = get_global_id(1);
         var3 = 1;
-        var2 = (uint)(get_global_id(1) * x) - (uint)y;
+        var2 = (get_global_id(1) * x) - (uint)y;
     }
     data[var3] = var2;
     data[var5] = x;
