+++ tests/branching_kernels/if_with_long_expr/if_with_long_expr_hands-gfx.cl
@@ -5,14 +5,14 @@ void if_with_long_expr(int x, __global int *data, int y)
     uint var2;
     int var4;
     var2 = get_global_id(0);
-    if ((int)x >= (int)y || (uint)1 != (uint)get_global_id(0) || (uint)2 != (uint)get_global_id(2)) {
+    if (x >= y || 1 != get_global_id(0) || 2 != get_global_id(2)) {
         var4 = get_global_id(0) * y;
         var1 = get_global_id(1);
     }
     else {
         var1 = get_global_id(1);
         var2 = 1;
-        var4 = (uint)(get_global_id(1) * x) - (uint)y;
+        var4 = (get_global_id(1) * x) - (uint)y;
     }
     data[var2] = var4;
     data[var1] = x;
