+++ tests/different_flags/flag_only_opencl/flag_only_opencl_dcmpl.cl
@@ -1,5 +1,5 @@
 Not resolved yet. flat_load_sshort v4, v[4:5]
 __kernel void sum(__global short *A, __global ulong *C, uint size, __local ulong *L)
 {
-    (uint)size > (uint)get_local_id(0);
+    size > (uint)get_local_id(0);
 }
