+++ tests/linear_kernels/many_linears/many_linears_dcmpl-gfx1030.cl
@@ -25,13 +25,13 @@ void copy4(int x, __global int *data)
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void copy5(int x, __global uint *data)
 {
-    data[get_global_id(0)] = (ulong)((ulong)((ulong)get_global_offset(2) + (ulong)get_global_offset(1)) + (ulong)(get_group_id(2) * 4)) + (ulong)(get_group_id(1) * 2) + (ulong)((ulong)get_local_id(2) + (ulong)get_local_id(1));
+    data[get_global_id(0)] = (ulong)(((ulong)get_global_offset(2) + (ulong)get_global_offset(1)) + (ulong)(get_group_id(2) * 4)) + (ulong)(get_group_id(1) * 2) + ((ulong)get_local_id(2) + (ulong)get_local_id(1));
 }
 
 __kernel __attribute__((reqd_work_group_size(2, 2, 4)))
 void copy6(int x, __global uint *data)
 {
-    data[get_global_id(0) - get_global_offset(0)] = (ulong)((ulong)((ulong)get_global_offset(2) + (ulong)get_global_offset(1)) + (ulong)(get_group_id(2) * 4)) + (ulong)(get_group_id(1) * 2) + (ulong)((ulong)get_local_id(2) + (ulong)get_local_id(1));
+    data[get_global_id(0) - get_global_offset(0)] = (ulong)(((ulong)get_global_offset(2) + (ulong)get_global_offset(1)) + (ulong)(get_group_id(2) * 4)) + (ulong)(get_group_id(1) * 2) + ((ulong)get_local_id(2) + (ulong)get_local_id(1));
 }
 
 __kernel __attribute__((reqd_work_group_size(64, 1, 1)))
