+++ tests/loops_kernels/loop_general_dec/loop_general_dec_dcmpl-gfx1010.cl
@@ -9,13 +9,13 @@ void loop_general(__global int *data, int from, int to, int x)
     var0 = data[get_global_id(0)];
     var4 = var0;
     var5 = from;
-    if ((int)var5 <= (int)to) {
+    if (var5 <= to) {
         do {
             var3 = x * var4;
-            var4 = (ulong)var5 + (ulong)(var3);
-            var5 = (ulong)var5 + (ulong)(-1);
-        } while (!((int)((ulong)var5 + (ulong)(-1)) <= (int)to));
-        var2 = (ulong)(var3) + (ulong)((ulong)var5 + (ulong)(-1)) + (ulong)1;
+            var4 = (ulong)var5 + (var3);
+            var5 = var5 + (ulong)(-1);
+        } while (!((var5 + (ulong)(-1)) <= to));
+        var2 = (var3) + (ulong)(var5 + (ulong)(-1)) + 1;
     }
     data[get_global_id(0)] = var2;
 }
