+++ tests/real_kernels/bellmanFordInit_ver1/bellmanFordInit_ver1_hands.cl
@@ -5,14 +5,14 @@ void bellmanFordIter_ver1(uint nEdges, __global uint2 *edges, __global double *w
     double var1;
     double var2;
     double var3;
-    if ((uint)nEdges > (uint)get_global_id(0)) {
+    if (nEdges > get_global_id(0)) {
         var0 = edges[get_global_id(0)];
         var1 = d[var0.s0];
-        if ((double)var1 < (double)1e12) {
+        if (var1 < (double)1e12) {
             var2 = d[var0.s1];
             var3 = weights[get_global_id(0)];
-            if ((double)var2 > (double)((double)var1 + (double)var3)) {
-                d[var0.s1] = (double)var1 + (double)var3;
+            if (var2 > (var1 + var3)) {
+                d[var0.s1] = var1 + var3;
                 *changed = 1;
             }
         }
